se agrega jakarta

+++ src/main/java/com/dh/Clase15_SpringMVC/servicio/ITurnoServicio.java
@@ -1,12 +1,12 @@
 package com.dh.Clase15_SpringMVC.servicio;
 
-import com.dh.Clase15_SpringMVC.modelo.Turno;
+import com.dh.Clase15_SpringMVC.entity.Turno;
 
 import java.util.List;
 
 public interface ITurnoServicio {
     //CRUD
     Turno guardar(Turno turno);
-    Turno buscarPorId(Integer id);
+    Turno buscarPorId(Long id);
     List<Turno> listarTodos();
 }
